Add logic from DFE for showing item table
